+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 🚀 Updates
 
+- Added "bundled" as a supported alias for `npm`.
 - Updated `node` to automatically install the `npm` version that comes bundled with Node.js.
 
 #### 🐞 Fixes